requirements.txt
requests
fastapi
fastapi-utils
uvicorn
pyTelegramBotAPI
python-dotenv
loguru
pytz
pickledb